requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.5
psycopg2-binary==2.8.5
requests==2.23.0
selenium==3.141.0
urllib3==1.25.7
yarl==1.4.2
nose==1.3.7
coverage==4.5.3
rednose==1.3.0
mock==3.0.5
freezegun==0.3.11
